Got Lights from Nodes
Not sure if this is the correct way, but I traversed the tree applying
the model view transform similar to the way he did for drawing the
bounding boxes (which includes the animation transform similar to how he
does). I started to implement the lights but ran into some problems on
how to do that similar to the way he did in his example. If you
uncomment the code in scenegraph.cpp, you will see why. You need to get
the number of lights from traversing the tree, but you also need the
shader program which is only currently available prior to using draw.
See what you think.